Success Message on View Component
Added functionality to the smart and dumb components from the actions. State now shows alert to user of succesfull registration as well as capabilities for deleting the message by deleting the id associated with the alert message. Now on to user authentication in the back end using Crypto for password authentication