remove property from assignment documents
It makes more sense to store the information of 'who an assignment is assigned to' on the gradebook document, because an assignment may get reused or shared for years, and only a gradebook would care who it is assigned to. We can therefore stop tracking this property on assignments, and start tracking it on gradebooks.